chore: validate LinterExclusions configuration
